refactor: Standardize votingEndTimestamp to uint48
Updates the `votingEndTimestamp` type to `uint48` across relevant contracts, including voting strategies, validators, Azorius, and mocks.

This change provides a significantly larger range for timestamps compared to `uint32` (preventing issues around the year 2106) while optimizing gas usage compared to `uint256`. The `currentTimestamp` in the `VotingPeriodEnded` event is also updated to `uint48` for consistency.

+++ contracts/deployables/strategies/ERC4337VoterSupportV1.sol
@@ -19,8 +19,8 @@ abstract contract ERC4337VoterSupportV1 is SmartAccountValidationV1 {
 
     event VotingPeriodEnded(
         uint32 indexed proposalId,
-        uint256 votingEndTimestamp,
-        uint256 currentTimestamp
+        uint48 votingEndTimestamp,
+        uint48 currentTimestamp
     );
 
     constructor() {
